+++ old-ui/app/components/account-dropdowns.js
@@ -2,13 +2,13 @@ const Component = require('react').Component
 const PropTypes = require('prop-types')
 const h = require('react-hyperscript')
 const actions = require('../../../ui/app/actions')
-const genAccountLink = require('etherscan-link').createAccountLink
 const connect = require('react-redux').connect
 const Dropdown = require('./dropdown').Dropdown
 const DropdownMenuItem = require('./dropdown').DropdownMenuItem
 const Identicon = require('./identicon')
 const ethUtil = require('ethereumjs-util')
 const copyToClipboard = require('copy-to-clipboard')
+const ethNetProps = require('eth-net-props')
 
 class AccountDropdowns extends Component {
   constructor (props) {
@@ -54,7 +54,7 @@ class AccountDropdowns extends Component {
             style: {
               width: '4px',
               height: '26px',
-              background: '#8fdc97',
+              background: '#60db97',
               position: 'absolute',
               left: '-25px',
             },
@@ -82,17 +82,32 @@ class AccountDropdowns extends Component {
             },
           }, identity.name || ''),
           this.indicateIfLoose(keyring),
+          this.ifLooseAcc(keyring) ? h('.remove', {
+            onClick: (event) => {
+              event.preventDefault()
+              event.stopPropagation()
+              this.props.actions.showDeleteImportedAccount(identity)
+              this.setState({
+                accountSelectorActive: false,
+                optionsMenuActive: false,
+              })
+            },
+          }) : null,
         ]
       )
     })
   }
 
-  indicateIfLoose (keyring) {
+  ifLooseAcc (keyring) {
     try { // Sometimes keyrings aren't loaded yet:
       const type = keyring.type
       const isLoose = type !== 'HD Key Tree'
-      return isLoose ? h('.keyring-label', 'IMPORTED') : null
+      return isLoose
     } catch (e) { return }
+  }
+
+  indicateIfLoose (keyring) {
+    return this.ifLooseAcc(keyring) ? h('.keyring-label', 'IMPORTED') : null
   }
 
   renderAccountSelector () {
@@ -104,12 +119,12 @@ class AccountDropdowns extends Component {
       {
         useCssTransition: true, // Hardcoded because account selector is temporarily in app-header
         style: {
-          marginLeft: '-198px',
+          marginLeft: '-213px',
           marginTop: '32px',
           minWidth: '180px',
           overflowY: 'auto',
           maxHeight: '300px',
-          width: '250px',
+          width: '265px',
         },
         innerStyle: {
           padding: '8px 25px',
@@ -135,7 +150,7 @@ class AccountDropdowns extends Component {
             onClick: () => actions.addNewAccount(),
           },
           [
-            h('span', { style: { fontSize: '16px', color: '#8fdc97' } }, 'Create Account'),
+            h('span', { style: { fontSize: '16px', color: '#60db97' } }, 'Create Account'),
           ],
         ),
         h(
@@ -152,26 +167,13 @@ class AccountDropdowns extends Component {
               style: {
                 fontSize: '16px',
                 marginBottom: '5px',
-                color: '#8fdc97',
+                color: '#60db97',
               },
             }, 'Import Account'),
           ]
         ),
       ]
     )
-  }
-
-  genPOAEplorerAccountLink (selected, network) {
-    const isSokol = parseInt(network) === 77
-    const isPOA = parseInt(network) === 99
-    if (isSokol) {
-      return `https://sokol.poaexplorer.com/address/search/${selected}`
-    }
-    if (isPOA) {
-      return `https://poaexplorer.com/address/search/${selected}`
-    }
-
-    return ''
   }
 
   renderAccountOptions () {
@@ -207,7 +209,7 @@ class AccountDropdowns extends Component {
             closeMenu: () => {},
             onClick: () => {
               const { selected, network } = this.props
-              const url = (isSokol || isPOA) ? this.genPOAEplorerAccountLink(selected, network) : genAccountLink(selected, network)
+              const url = ethNetProps.explorerLinks.getExplorerAccountLinkFor(selected, network)
               global.platform.openWindow({ url })
             },
           },
@@ -332,6 +334,7 @@ const mapDispatchToProps = (dispatch) => {
       addNewAccount: () => dispatch(actions.addNewAccount()),
       showImportPage: () => dispatch(actions.showImportPage()),
       showQrView: (selected, identity) => dispatch(actions.showQrView(selected, identity)),
+      showDeleteImportedAccount: (identity) => dispatch(actions.showDeleteImportedAccount(identity)),
     },
   }
 }