+++ old-ui/app/components/token-cell.js
@@ -2,7 +2,7 @@ const Component = require('react').Component
 const h = require('react-hyperscript')
 const inherits = require('util').inherits
 const Identicon = require('./identicon')
-const prefixForNetwork = require('../../lib/etherscan-prefix-for-network')
+const ethNetProps = require('eth-net-props')
 
 module.exports = TokenCell
 
@@ -49,8 +49,6 @@ TokenCell.prototype.render = function () {
         },
       }, ''),
 
-      h('hr'),
-
       /*
       h('button', {
         onClick: this.send.bind(this, address),
@@ -71,7 +69,7 @@ TokenCell.prototype.send = function (address, event) {
 }
 
 TokenCell.prototype.view = function (address, userAddress, network, event) {
-  const url = etherscanLinkFor(address, userAddress, network)
+  const url = ethNetProps.explorerLinks.getExplorerTokenLinkFor(address, userAddress, network)
   if (url) {
     navigateTo(url)
   }
@@ -79,11 +77,6 @@ TokenCell.prototype.view = function (address, userAddress, network, event) {
 
 function navigateTo (url) {
   global.platform.openWindow({ url })
-}
-
-function etherscanLinkFor (tokenAddress, address, network) {
-  const prefix = prefixForNetwork(network)
-  return `https://${prefix}etherscan.io/token/${tokenAddress}?a=${address}`
 }
 
 function tokenFactoryFor (tokenAddress) {