+++ old-ui/app/first-time/init-menu.js
@@ -77,7 +77,7 @@ InitializeMenuScreen.prototype.renderMenu = function (state) {
             style: {
               fontSize: '18px',
               position: 'relative',
-              color: '#8fdc97',
+              color: '#60db97',
               top: '2px',
               marginLeft: '4px',
             },
@@ -134,7 +134,7 @@ InitializeMenuScreen.prototype.renderMenu = function (state) {
           onClick: this.showRestoreVault.bind(this),
           style: {
             fontSize: '0.8em',
-            color: '#8fdc97',
+            color: '#60db97',
           },
         }, 'Import Existing DEN'),
       ]),