+++ test/e2e/func.js
@@ -5,7 +5,8 @@ const os = require('os')
 const path = require('path')
 const webdriver = require('selenium-webdriver')
 const Command = require('selenium-webdriver/lib/command').Command
-const By = webdriver.By
+
+const { By, Key } = webdriver
 
 module.exports = {
   delay,
@@ -14,7 +15,8 @@ module.exports = {
   installWebExt,
   getExtensionIdChrome,
   getExtensionIdFirefox,
-}
+  clearField,
+  }
 
 function delay (time) {
   return new Promise(resolve => setTimeout(resolve, time))
@@ -60,4 +62,12 @@ async function installWebExt (driver, extension) {
     .defineCommand(cmd.getName(), 'POST', '/session/:sessionId/moz/addon/install')
 
   return await driver.schedule(cmd, 'installWebExt(' + extension + ')')
+}
+
+async function clearField (field, number) {
+  await field.click()
+  if (number === undefined) number = 40
+  for (let i = 0; i < number; i++) {
+    await field.sendKeys(Key.BACK_SPACE)
+  }
 }