+++ ui/app/components/modals/account-details-modal.js
@@ -6,7 +6,7 @@ const connect = require('react-redux').connect
 const actions = require('../../actions')
 const AccountModalContainer = require('./account-modal-container')
 const { getSelectedIdentity } = require('../../selectors')
-const genAccountLink = require('../../../lib/account-link.js')
+const ethNetProps = require('eth-net-props')
 const QrView = require('../qr-code')
 const EditableLabel = require('../editable-label')
 
@@ -69,7 +69,7 @@ AccountDetailsModal.prototype.render = function () {
       h('div.account-modal-divider'),
 
       h('button.btn-primary.account-modal__button', {
-        onClick: () => global.platform.openWindow({ url: genAccountLink(address, network) }),
+        onClick: () => global.platform.openWindow({ url: ethNetProps.explorerLinks.getExplorerAccountLinkFor(address, network) }),
       }, this.context.t('etherscanView')),
 
       // Holding on redesign for Export Private Key functionality